backend/requirements.txt
djangorestframework-simplejwt
requests
asgiref==3.10.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.7
django-rest-framework==0.1.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
Werkzeug==3.1.3
